Put CustomAutoSchema in views instead of new file and fixed other things
